requirements.txt
blinker==1.6.2
click==8.1.7
Flask==2.3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.25.2
Pillow==10.0.1
redis==5.0.0
pyuWSGI==2.0.22
Werkzeug==2.3.7
